% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{addClass}
\alias{addClass}
\title{Adds cls to class(x).}
\usage{
addClass(x, cls)
}
\arguments{
\item{x}{an object to add the class cls to}

\item{cls}{a class}
}
\value{
the x argument
}
\description{
Adds cls to class(x).
}
