% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{archF}
\alias{archF}
\title{Moves the file to a proper archive/ subdir}
\usage{
archF(file)
}
\arguments{
\item{file}{path/to/some/file.ext}
}
\value{
path/to/some/archive/file.ext
}
\description{
Moves the file to a proper archive/ subdir
}
