% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{asNonBlanks}
\alias{asNonBlanks}
\title{Explicitly marks given strings as trimmed and non-blank.
User takes responsibility.}
\usage{
asNonBlanks(s)
}
\arguments{
\item{s}{a string to mark as trimmed}
}
\value{
s
}
\description{
Explicitly marks given strings as trimmed and non-blank.
User takes responsibility.
}
