% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{assocKvs}
\alias{assocKvs}
\title{Puts all key/value mapping pairs into m.}
\usage{
assocKvs(m, kvs)
}
\arguments{
\item{m}{associative container, list or env}

\item{kvs}{list of key/value pairs (must be even in size)}
}
\value{
updated m
}
\description{
Puts all key/value mapping pairs into m.
}
