% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{asTrimmed}
\alias{asTrimmed}
\title{Explicitly marks given strings as trimmed. User takes responsibility.}
\usage{
asTrimmed(s)
}
\arguments{
\item{s}{a string to mark as trimmed}
}
\value{
s
}
\description{
Explicitly marks given strings as trimmed. User takes responsibility.
}
