% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bindDTs}
\alias{bindDTs}
\title{Binds the data.table arguments into one data.table, optionally filling}
\usage{
bindDTs(..., fill = FALSE)
}
\value{
the resulting data.table
}
\description{
Binds the data.table arguments into one data.table, optionally filling
}
