% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{boolsAnd}
\alias{boolsAnd}
\title{Short-circuit evaluator for Boolean values}
\usage{
boolsAnd(b, x, pred)
}
\arguments{
\item{b}{a boolean vector of some length n}

\item{x}{a vector of lenght presumably equal to n}

\item{pred}{a predicate to evaluate on these elements of x for which b is TRUE}
}
\value{
a boolean vector of length n
}
\description{
Short-circuit evaluator for Boolean values
}
