% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{brewerTheme}
\alias{brewerTheme}
\title{Custom pleasant theme from: http://minimaxir.com/2015/02/ggplot-tutorial/}
\usage{
brewerTheme(paletteName = "Greys", textSize = 7)
}
\description{
Custom pleasant theme from: http://minimaxir.com/2015/02/ggplot-tutorial/
}

