% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{captArgs}
\alias{captArgs}
\title{Captures the arguments for calling f. When called without f returns
the last captured args for given id.}
\usage{
captArgs(id, f = NULL, ...)
}
\description{
Captures the arguments for calling f. When called without f returns
the last captured args for given id.
}
