% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{catimela}
\alias{catimela}
\title{\code{cat}s (logs) time elapsed for expr to evaluate}
\usage{
catimela(expr, msg = "", off = FALSE)
}
\arguments{
\item{expr}{expression to be evaluated}

\item{msg}{optional message for easier identification of what was evaluated}

\item{off}{optional flag for turning off when needed}
}
\value{
the value of evaluated expr
}
\description{
\code{cat}s (logs) time elapsed for expr to evaluate
}
