% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{chatty}
\alias{chatty}
\title{Debug wrapper generator.}
\usage{
chatty(f, prefix = "Processing ")
}
\arguments{
\item{f}{Function to wrap}

\item{prefix}{A prefix for the diagnostic message}
}
\description{
Returns a diagnostic wrapper around f. Thanks H. Wickham.
}
