% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{colNamed}
\alias{colNamed}
\title{By name column selection.}
\usage{
colNamed(name, df = NULL)
}
\arguments{
\item{name}{The column name}

\item{df}{Data frame or data table}
}
\value{
The column as a vector
}
\description{
\code{colNamed} when called with a character name and a data frame or a
data table returns its named column. When called with name only returns a
corresponding clojure having a df argument.
}

