% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copyDoubles}
\alias{copyDoubles}
\title{Works like System.arraycopy(...) in Java for double vectors.}
\usage{
copyDoubles(src, srcPos, dest, destPos, length)
}
\description{
Works like System.arraycopy(...) in Java for double vectors.
}
