% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copyInts}
\alias{copyInts}
\title{Works like System.arraycopy(...) in Java for integer vectors.}
\usage{
copyInts(src, srcPos, dest, destPos, length)
}
\description{
Works like System.arraycopy(...) in Java for integer vectors.
}
