% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{countMatching}
\alias{countMatching}
\title{Reports the number of elements matching the predicate.}
\usage{
countMatching(.data, pred, ...)
}
\arguments{
\item{.data}{The data object (vector, frame).}

\item{pred}{The predicate.}

\item{keepZeros}{For non-atomics. When TRUE the zero counts will be
kept.}
}
\value{
A representation of counts.
}
\description{
Reports the number of elements matching the predicate.
}

