% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countWorkdays}
\alias{countWorkdays}
\title{Computes number of work days (1-5/Mon-Fri) fot the total number of days and
starting with the initial day of week (1-7/Mon-Sun).}
\usage{
countWorkdays(dayOfWeek, daysCount)
}
\description{
Computes number of work days (1-5/Mon-Fri) fot the total number of days and
starting with the initial day of week (1-7/Mon-Sun).
}
