% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{createEnvMap}
\alias{createEnvMap}
\title{Creates an env using the enumerated collection of key/value mapping pairs}
\usage{
createEnvMap(...)
}
\arguments{
\item{...}{collection of key/value pairs (must be even in size)}
}
\value{
the resulting env
}
\description{
Creates an env using the enumerated collection of key/value mapping pairs
}
