% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{createListMap}
\alias{createListMap}
\title{Creates a list using the enumerated collection of key/value mapping pairs.}
\usage{
createListMap(...)
}
\arguments{
\item{...}{collection of key/value pairs (must be even in size)}
}
\value{
the resulting list
}
\description{
Creates a list using the enumerated collection of key/value mapping pairs.
}
