% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{disj}
\alias{disj}
\title{Peformance optimized base::setdiff}
\usage{
disj(x, y)
}
\arguments{
\item{x}{an input vector}

\item{y}{vector of elements to filter out from x}
}
\value{
x without elements of y
}
\description{
Peformance optimized base::setdiff
}
