% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{doCall}
\alias{doCall}
\title{Works like \code{do.call} but optionaly removes all NULL args.}
\usage{
doCall(what, args, quote = FALSE, envir = parent.frame(),
  null.rm = TRUE)
}
\description{
Works like \code{do.call} but optionaly removes all NULL args.
}
