% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{duplicatedBy}
\alias{duplicatedBy}
\title{Determine and return duplicated rows.}
\usage{
duplicatedBy(x, by, ...)
}
\description{
\code{duplicatedBy} returns rows having same values in the specified by col.
}

