% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{fte_theme}
\alias{fte_theme}
\title{Custom pleasant theme from: http://minimaxir.com/2015/02/ggplot-tutorial/}
\usage{
fte_theme(brewerPaletteName = "Greys")
}
\description{
Custom pleasant theme from: http://minimaxir.com/2015/02/ggplot-tutorial/
}

