% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDTcolsMatching}
\alias{getDTcolsMatching}
\title{Returns a set of dt columns c such that match predicate pred on values.}
\usage{
getDTcolsMatching(dt, pred, quant = any)
}
\arguments{
\item{dt}{a data.table}

\item{pred}{a function}

\item{quant}{rows-related quantifier function, e.g. any, every}
}
\value{
a resulting set of columns
}
\description{
Returns a set of dt columns c such that match predicate pred on values.
}
