% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{hashmap}
\alias{hashmap}
\title{Creates an empty hashmap (environment).}
\usage{
hashmap(size = NA)
}
\arguments{
\item{size}{See new.env documentation}
}
\description{
Creates an empty hashmap (environment).
}

