% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{listDiscard}
\alias{listDiscard}
\title{Removes (discards) elements in list l for which p is TRUE.}
\usage{
listDiscard(l, p)
}
\arguments{
\item{l}{a list}

\item{p}{a predicate}
}
\value{
a resulting list
}
\description{
Removes (discards) elements in list l for which p is TRUE.
}
