% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{listKeep}
\alias{listKeep}
\title{Keeps elements in list l for which p is TRUE.}
\usage{
listKeep(l, p)
}
\arguments{
\item{l}{a list}

\item{p}{a predicate}
}
\value{
a resulting list
}
\description{
Keeps elements in list l for which p is TRUE.
}
