% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mapDT}
\alias{mapDT}
\title{Maps f over each singleton data.tables representing the rows of dt}
\usage{
mapDT(dt, f)
}
\arguments{
\item{dt}{chDT}

\item{f}{chFun chDT1 → chDT}
}
\value{
chDT
}
\description{
Maps f over each singleton data.tables representing the rows of dt
}
