% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memoBool}
\alias{memoBool}
\title{A fast memoizer for values of function call depending on some logical
condition.}
\usage{
memoBool(f, ...)
}
\description{
A fast memoizer for values of function call depending on some logical
condition.
}
