% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{modes}
\alias{modes}
\title{Returns a data.table of mode values, together with their frequency.}
\usage{
modes(x)
}
\description{
Returns a data.table of mode values, together with their frequency.
}

