% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{mostFrequent}
\alias{mostFrequent}
\title{Returns a data-frame of n most frequent elements in vector x
together with their frequencies.}
\usage{
mostFrequent(x, n = 10)
}
\description{
Returns a data-frame of n most frequent elements in vector x
together with their frequencies.
}

