% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{moveDTcols}
\alias{moveDTcols}
\title{Uses \code{moveNames} to set a new column ordering (destructive on dt)}
\usage{
moveDTcols(dt, ...)
}
\description{
Uses \code{moveNames} to set a new column ordering (destructive on dt)
}
