% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{moveDTprops__}
\alias{moveDTprops__}
\title{Diagnostic version of \code{moveDTprops}}
\usage{
moveDTprops__(dt, ...)
}
\arguments{
\item{dt}{chDT|chR(chDT)}
}
\value{
the dt argument
}
\description{
Diagnostic version of \code{moveDTprops}
}
