% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{nonBlanks}
\alias{nonBlanks}
\title{Asserts non-blankness and returns koR.Trimmed strings as NonBlank.}
\usage{
nonBlanks(s)
}
\arguments{
\item{s}{koR.Trimmed strings}
}
\value{
koR.NonBlank strings
}
\description{
Asserts non-blankness and returns koR.Trimmed strings as NonBlank.
}
