% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ofClass}
\alias{ofClass}
\title{Sets class(x) to cls.}
\usage{
ofClass(x, cls)
}
\arguments{
\item{x}{an object to set the class c to}

\item{cls}{a class}
}
\value{
the x argument with class(x) set to cls
}
\description{
Sets class(x) to cls.
}
