% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{overDT}
\alias{overDT}
\title{Sets data.table's new value of a column to a value of a function call
with an old value of the column. Does not check if column exists in dt.}
\usage{
overDT(dt, col, f)
}
\arguments{
\item{dt}{a data.table}

\item{col}{a column of dt}

\item{f}{a function}
}
\value{
dt
}
\description{
Sets data.table's new value of a column to a value of a function call
with an old value of the column. Does not check if column exists in dt.
}
