% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{partition}
\alias{partition}
\title{Splits xs into partitions of length n, thanks to
https://stackoverflow.com/questions/3318333/split-a-vector-into-chunks-in-r}
\usage{
partition(xs, n)
}
\arguments{
\item{xs}{an atomic vector to partition}

\item{n}{the partition size}
}
\value{
a list containing the partitions
}
\description{
Splits xs into partitions of length n, thanks to
https://stackoverflow.com/questions/3318333/split-a-vector-into-chunks-in-r
}
