% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/df.R
\name{printColsCounts}
\alias{printColsCounts}
\title{Report matching column values counts.}
\usage{
printColsCounts(df, pred, printZeros = FALSE)
}
\arguments{
\item{df}{The data frame}

\item{pred}{The predicate}

\item{printZeros}{When TRUE the zero counts will be printed out}
}
\description{
\code{printColsCounts} prints counts of items in data frame columns that
match the given predicate.
}

