% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{printCondition}
\alias{printCondition}
\title{Condition handler generator.}
\usage{
printCondition(prefix = "ERROR: ")
}
\arguments{
\item{prefix}{Prefix for the condition message.}
}
\value{
A condition handler.
}
\description{
\code{printCondition} returns a condition handler that prints a
prefixed conditionMessage.
}

