% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qbench}
\alias{qbench}
\title{Alias for \code{microbenchmark::microbenchmark}}
\usage{
qbench(..., list = NULL, times = 100L, unit, check = NULL,
  control = list())
}
\description{
Alias for \code{microbenchmark::microbenchmark}
}
