% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{savePlot}
\alias{savePlot}
\title{Saves a plot using a graphics device and returns the plot.}
\usage{
savePlot(p, grDevice, ...)
}
\arguments{
\item{p}{(not only) ggplot2 plot}

\item{grDevice}{graphics device}

\item{...}{arguments to grDevice(...)}
}
\value{
p
}
\description{
Saves a plot using a graphics device and returns the plot.
}
