% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{setClass}
\alias{setClass}
\title{Sets class(x) to c and returns x}
\usage{
setClass(x, c)
}
\arguments{
\item{x}{an object to set the class c to}

\item{c}{a class}
}
\value{
the x argument
}
\description{
Sets class(x) to c and returns x
}
