% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setDT}
\alias{setDT}
\title{Sets data.table's column value to value and returns the data.table.}
\usage{
setDT(dt, col, value)
}
\arguments{
\item{dt}{a data.table}

\item{col}{a column of dt}

\item{value}{to set}
}
\value{
dt
}
\description{
Sets data.table's column value to value and returns the data.table.
}
