% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{setDTpropsorder__}
\alias{setDTpropsorder__}
\title{Diagnostic version of \code{setDTpropsorder}}
\usage{
setDTpropsorder__(dt, neworder)
}
\arguments{
\item{dt}{chDT|chR(chDT)}
}
\value{
the dt argument
}
\description{
Diagnostic version of \code{setDTpropsorder}
}
