% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{settingClass}
\alias{settingClass}
\title{Sets class(x) to c and returns x.}
\usage{
settingClass(x, c)
}
\arguments{
\item{x}{an object to set the class c to}

\item{c}{a class}
}
\value{
the x argument
}
\description{
Sets class(x) to c and returns x.
}
