% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{skewness}
\alias{skewness}
\title{A wrapper around \code{skewnessRcpp} safe with respect to NAs.}
\usage{
skewness(x, na.rm = FALSE)
}
\description{
A wrapper around \code{skewnessRcpp} safe with respect to NAs.
}

