% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skewnessRcpp}
\alias{skewnessRcpp}
\title{Returns the skewness of the vector. No NAs assumed.}
\usage{
skewnessRcpp(x)
}
\arguments{
\item{x}{the vector}
}
\description{
Returns the skewness of the vector. No NAs assumed.
}

