% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{stampString}
\alias{stampString}
\title{Returns a timestamp string with the tzone appended.}
\usage{
stampString(tzone = "CEST", hsep = "_")
}
\description{
Returns a timestamp string with the tzone appended.
}
