% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{strInsertAfterIndex}
\alias{strInsertAfterIndex}
\title{Inserts what after a non-negative index i in string s. Vectorized over all args.}
\usage{
strInsertAfterIndex(s, i, what)
}
\arguments{
\item{s}{a string to transform}

\item{i}{a non-negative index}

\item{what}{to insert}
}
\value{
a transformed string
}
\description{
Inserts what after a non-negative index i in string s. Vectorized over all args.
}
