% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{strInsertBeforeIndex}
\alias{strInsertBeforeIndex}
\title{Inserts what before positive index i in string s. Vectorized over all args.}
\usage{
strInsertBeforeIndex(s, i, what)
}
\arguments{
\item{s}{a string to transform}

\item{i}{a positive index}

\item{what}{to insert}
}
\value{
a transformed string
}
\description{
Inserts what before positive index i in string s. Vectorized over all args.
}
