% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{strWrapAll}
\alias{strWrapAll}
\title{Wraps every occurence of the pattern with a prefix and a suffix. Non-vectorized,
the arguments are expected to be scalars (atoms).}
\usage{
strWrapAll(s, pattern, prefix, suffix, ignoreCase = TRUE)
}
\arguments{
\item{s}{a string to transform}

\item{pattern}{describing the pieces of s to wrap}

\item{prefix}{: chString}

\item{suffix}{: chString}

\item{ignoreCase}{: chBool}
}
\value{
a transformed string
}
\description{
Wraps every occurence of the pattern with a prefix and a suffix. Non-vectorized,
the arguments are expected to be scalars (atoms).
}
