% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sysSleeping}
\alias{sysSleeping}
\title{Returns a decorator of f that sleeps time (in seconds) before executing f.}
\usage{
sysSleeping(f, time)
}
\description{
Returns a decorator of f that sleeps time (in seconds) before executing f.
}
