% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{timeit}
\alias{timeit}
\title{Alias for \code{microbenchmark::microbenchmark}}
\usage{
timeit(..., list = NULL, times = 100L, unit, check = NULL,
  control = list())
}
\description{
Alias for \code{microbenchmark::microbenchmark}
}

