% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{trimStrings}
\alias{trimStrings}
\title{Trims given strings (both ends)}
\usage{
trimStrings(s)
}
\arguments{
\item{s}{a vector of strings to trim}
}
\value{
a vector of trimmed strings
}
\description{
Trims given strings (both ends)
}
