% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{userInstalledPackages}
\alias{userInstalledPackages}
\title{Gets user installed packages.}
\usage{
userInstalledPackages()
}
\value{
a data table containing all the relevant information
}
\description{
Gets user installed packages.
}
\seealso{
installed.packages()
}

