% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{withDTcols}
\alias{withDTcols}
\title{Builds a data.table by using only cols of dt}
\usage{
withDTcols(dt, cols)
}
\arguments{
\item{dt}{a data.table}

\item{cols}{a vector of Strings}
}
\description{
Builds a data.table by using only cols of dt
}
