% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.R
\name{withoutDTcols}
\alias{withoutDTcols}
\title{Builds a data.table by removing cols of dt}
\usage{
withoutDTcols(dt, cols)
}
\arguments{
\item{dt}{a data.table}

\item{cols}{a vector of Strings}
}
\description{
Builds a data.table by removing cols of dt
}
