% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{withStr}
\alias{withStr}
\title{Takes a vector of Strings and returns a new vector with a new String s2 moving
it onto a pos related to s1.}
\usage{
withStr(s, s2, pos = "after", s1)
}
\arguments{
\item{s}{a vector of Strings}

\item{s2}{a new String}

\item{pos}{see \code{moveNames}}

\item{s1}{an element in s according to which we position s2}
}
\value{
a new vector of Strings with s2 pos(itioned) according to s1
}
\description{
Takes a vector of Strings and returns a new vector with a new String s2 moving
it onto a pos related to s1.
}
