% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.R
\name{writeReportFile}
\alias{writeReportFile}
\title{Generates a report file for data.table dt in a particular year}
\usage{
writeReportFile(dt, year = NULL, problem = NULL)
}
\description{
Generates a report file for data.table dt in a particular year
}
