% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{addClass}
\alias{addClass}
\title{Adds c to class(x) and returns x.}
\usage{
addClass(x, c)
}
\arguments{
\item{x}{an object to add the class c to}

\item{c}{a class}
}
\value{
the x argument
}
\description{
Adds c to class(x) and returns x.
}
