% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{ensureF}
\alias{ensureF}
\title{Ensures that either file argument or its archived version exist}
\usage{
ensureF(file, archFile = NULL)
}
\arguments{
\item{file}{path/to/some/file.ext}
}
\value{
path/to/some/file.ext or archive/path/to/some/file.ext
}
\description{
Ensures that either file argument or its archived version exist
}
