% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memoBool}
\alias{memoBool}
\title{A fast memoizer for values of depending on some logical condition.}
\usage{
memoBool(forTrue, forFalse)
}
\description{
A fast memoizer for values of depending on some logical condition.
}
