% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mostFrequent}
\alias{mostFrequent}
\title{Returns a vector xs sorted by the (decreasing) frequency of appearance of
its elements.}
\usage{
mostFrequent(xs, ...)
}
\description{
Returns a vector xs sorted by the (decreasing) frequency of appearance of
its elements.
}
