% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{moveNames}
\alias{moveNames}
\title{Thanks to: https://stackoverflow.com/questions/18339370/reordering-columns-in-a-large-dataframe}
\usage{
moveNames(names, tomove, pos = "last", what = NULL)
}
\description{
Thanks to: https://stackoverflow.com/questions/18339370/reordering-columns-in-a-large-dataframe
}
