% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rgc}
\alias{rgc}
\title{Execute gc multiple times.}
\usage{
rgc(n = 10)
}
\arguments{
\item{n}{Number of calls.}
}
\value{
The result of the last gc() call.
}
\description{
\code{rgc} Calls gc() n times, 1 when n is negative.
}

