% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{setDTpropsorder__}
\alias{setDTpropsorder__}
\title{Diagnostic version of \code{setDTpropsorder}}
\usage{
setDTpropsorder__(dt, neworder)
}
\value{
chDT
}
\description{
Diagnostic version of \code{setDTpropsorder}
}
